partially updated MUI design

+++ frontend/src/components/SinglePost.jsx
@@ -1,10 +1,26 @@
+import { Button } from "@mui/material"
+import EditModal from "../forms/EditModal"
+import { useUserContext } from "../context/UserContext"
+
 function SinglePost(post) {
+    const {currentUser} = useUserContext()
+
     const deletePost = () => {
-        post.onDeletePost(post._id)
+        if (currentUser.username === post.postUsername) {
+            post.onDeletePost(post._id)
+        } else {
+            alert("Cannot delete posts that you did not create")
+            console.log('Failed to delete post: username does not match')
+        }
     }
 
     const updatePost = () => {
-        post.onUpdatePost(post._id)
+        if (currentUser.username === post.postUsername) {
+            post.onUpdatePost(post._id)
+        } else {
+            alert("Cannot edit posts that you did not create")
+            console.log('Failed to edit post: username does not match')
+        }
     }
 
     return (
@@ -12,15 +28,24 @@ function SinglePost(post) {
             <section className="border margin card">
                 <figure>
                     <div className="flex margin" style={{justifyContent: "space-between", alignItems: "center"}}>
-                        <h2>{post.title}</h2>
-                        <div><button onClick={() => updatePost(post)}>Edit</button>
-                            <button onClick={() => deletePost(post)}>Delete</button></div>
+                        <section className="flex padding align-left">
+                            {/* <img alt="avatar" src="https://placehold.co/70x70" style={{borderRadius: "50%"}} /> */}
+                            <h2>{post.title}</h2>
+                        </section>
+                        <div>
+                            <EditModal />
+                            <Button variant="outlined" style={{width: "6em"}} onClick={() => deletePost(post)}>Delete</Button>
+                        </div>
                     </div>
                     <hr />
                     <img className="margin" alt={post.alt} src={post.image} style={{maxWidth: "90%"}} />
-                    <figcaption className="align-left margin">{post.caption}</figcaption>
+                    <hr />
+                    <section className="margin align-left">
+                        <p><strong>By {post.postUsername}</strong></p>
+                        <div>{post.caption}</div>
+                    </section>
                 </figure>
-                <figure>
+                {/* <figure>
                     <div className="flex margin border padding">
                         <img alt="avatar" src="https://placehold.co/70x70" style={{borderRadius: "50%"}} />
                         <section className="padding align-left">
@@ -28,7 +53,7 @@ function SinglePost(post) {
                             <div>This is a comment about this comic strip</div>
                         </section>
                     </div>
-                </figure>
+                </figure> */}
             </section>
         </div>
     )
